Addresses issue #5434 on inconsistent selection of reads for PL AD and DP calculations.
It turns out that we use a wired 2bp overlap values for PL but 0 for AD. As a result some variants
with convinicing PL have 0 AD to support alternative. This change make sure that the same radius is used for
both and give the option to change it using an additional argument in HC and Mutec:

--allele-informative-reads-overlap-radius INT (default == 2).

We keep the wired value for the PL calculation since that affects sensitivity and we can observed a fall in QUAL values in integration test alone.
However, this is not compeling evidence that indeed this is the most appropriate value and a full evaluation is in order.

Now the AD and DP annotations would take in consideration this new radius (was 0 before). This will result in large values for AD and DP. In practice
DP nearly every time goes up by a few reads whereas AD does it more rarely. This is due to the fact that the radius increase often just add reads
that are actually not informative for PL and so the have the same likelihood for each variants and they don't contribute to AD.

When AD increases though one may thing that that is a good thing as we are adding additional suportive reads.... however you need to notice that
the `why we simply don't use all the reads in the active region in the first place (radius = +Infinity) this could often apply here... these reads are providing
evidence thru linakge-desecilibrium (arguably a good thing) or perhaps a defficient assembly with missing key haplotypes (bad for sure).

This makes is more likely that sum(AD) diverges from DP which some user may not be happy (although is not incorrect) but at least non 0,0,0 PL will now in hand with non 0
AD and DPs which some user may find more appropriate.

In anycase the question still remains on whether a radius of 2 is ideal.